requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Task Queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# AI/ML
openai==1.3.7
transformers==4.36.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Search Engine
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# Caching
redis==5.0.1
aioredis==2.0.1

# File Storage
boto3==1.34.0
minio==7.2.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Email
fastapi-mail==1.4.1
aiosmtplib==3.0.1

# SMS
twilio==8.10.3

# Social Media APIs
facebook-sdk==3.1.0
tweepy==4.14.0
instagram-basic-display==1.0.0

# Google Services
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
googlemaps==4.10.0

# Payment Processing
stripe==7.8.0
paypalrestsdk==1.13.3

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Development
python-dotenv==1.0.0
watchfiles==0.21.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.1.0

# Background Tasks
apscheduler==3.10.4

# WebSocket
websockets==12.0

# Data Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Compression
gzip==0.0.1

# Environment
python-decouple==3.8
